sync contract examples with other SDKs
